requirements.txt
torch==2.7.0
torchvision==0.22.0
open-clip-torch==2.32.0

numpy==2.0.2
scipy==1.13.1
scikit-learn==1.6.1
pandas==2.2.3
matplotlib==3.9.4

Pillow==11.2.1
PyYAML==6.0.2
tqdm==4.67.1
requests==2.32.3
ftfy==6.3.1
regex==2024.11.6
huggingface-hub==0.31.0
timm==1.0.15
safetensors==0.5.3
wandb==0.19.10

# Optional (needed for grad_sign/finetune_openCLIP.py)
lightning
torchmetrics
clip @ git+https://github.com/openai/CLIP.git
